allow selecting objects while in play mode
