+++ apps/client/src/editor/components/InspectMenu/InspectMenu.tsx
@@ -1,35 +1,16 @@
-import { ColliderExtension, SPAWN_TITLE } from "engine";
-import { nanoid } from "nanoid";
-import { useState } from "react";
-
-import Button from "../../../ui/Button";
-import {
-  DropdownContent,
-  DropdownItem,
-  DropdownMenu,
-  DropdownMenuItemProps,
-  DropdownTrigger,
-} from "../../../ui/DropdownMenu";
 import { useNode } from "../../hooks/useNode";
 import { useNodeAttribute } from "../../hooks/useNodeAttribute";
 import { useSpawn } from "../../hooks/useSpawn";
 import { useEditorStore } from "../../store";
+import AddComponentButton, { COMPONENT_TYPE, ComponentType } from "./AddComponentButton";
 import MeshComponent from "./mesh/MeshComponent";
 import PhysicsComponent from "./PhysicsComponent";
 import ScriptComponent from "./ScriptComponent";
 import SpawnPointComponent from "./SpawnPointComponent";
 import TransformComponent from "./TransformComponent";
 
-const COMPONENT_TYPE = {
-  Mesh: "Mesh",
-  Physics: "Physics",
-  SpawnPoint: "Spawn Point",
-  Script: "Script",
-} as const;
-
-type ComponentType = (typeof COMPONENT_TYPE)[keyof typeof COMPONENT_TYPE];
-
 export default function InspectMenu() {
+  const isPlaying = useEditorStore((state) => state.isPlaying);
   const selectedId = useEditorStore((state) => state.selectedId);
   const node = useNode(selectedId);
   const name = useNodeAttribute(selectedId, "name");
@@ -37,8 +18,6 @@ export default function InspectMenu() {
   const extensions = useNodeAttribute(selectedId, "extensions");
   const extras = useNodeAttribute(selectedId, "extras");
   const spawn = useSpawn();
-
-  const [open, setOpen] = useState(false);
 
   if (!node || !selectedId) return null;
 
@@ -54,10 +33,11 @@ export default function InspectMenu() {
         <input
           type="text"
           value={name ?? ""}
-          onChange={(e) => {
-            node.setName(e.target.value);
-          }}
-          className="mx-10 w-full rounded-lg py-0.5 text-center text-2xl font-bold transition hover:bg-neutral-200/80 focus:bg-neutral-200/80"
+          disabled={isPlaying}
+          onChange={(e) => node.setName(e.target.value)}
+          className={`mx-10 w-full rounded-lg py-0.5 text-center text-2xl font-bold transition ${
+            isPlaying ? "disabled:bg-white" : "hover:bg-neutral-200/80 focus:bg-neutral-200/80"
+          }`}
         />
       </div>
 
@@ -73,104 +53,13 @@ export default function InspectMenu() {
         })}
 
         {availableComponents.length > 0 && (
-          <div className="flex w-full justify-center">
-            <DropdownMenu open={open} onOpenChange={setOpen}>
-              <DropdownTrigger asChild>
-                <Button className="rounded-lg px-8">Add Component</Button>
-              </DropdownTrigger>
-
-              <DropdownContent open={open}>
-                <div className="py-2">
-                  {availableComponents.includes("Mesh") && (
-                    <ComponentButton
-                      onClick={() => {
-                        const { engine } = useEditorStore.getState();
-                        if (!engine) return;
-
-                        const { object: mesh } = engine.scene.mesh.create({
-                          extras: {
-                            customMesh: {
-                              type: "Box",
-                              width: 1,
-                              height: 1,
-                              depth: 1,
-                            },
-                          },
-                        });
-
-                        mesh.setName(node.getName());
-                        node.setMesh(mesh);
-                      }}
-                    >
-                      {COMPONENT_TYPE.Mesh}
-                    </ComponentButton>
-                  )}
-
-                  {availableComponents.includes(COMPONENT_TYPE.Physics) && (
-                    <ComponentButton
-                      onClick={() => {
-                        const { engine } = useEditorStore.getState();
-                        if (!engine) return;
-
-                        const collider = engine.scene.extensions.collider.createCollider();
-                        collider.type = "trimesh";
-
-                        node.setExtension(ColliderExtension.EXTENSION_NAME, collider);
-                      }}
-                    >
-                      {COMPONENT_TYPE.Physics}
-                    </ComponentButton>
-                  )}
-
-                  {availableComponents.includes(COMPONENT_TYPE.Script) && (
-                    <ComponentButton
-                      onClick={() => {
-                        const { engine } = useEditorStore.getState();
-                        if (!engine || !extras) return;
-
-                        const newExtras = { ...extras };
-                        if (!newExtras.scripts) newExtras.scripts = [];
-
-                        newExtras.scripts.push({ id: nanoid(), name: "New Script" });
-
-                        node.setExtras(newExtras);
-                      }}
-                    >
-                      {COMPONENT_TYPE.Script}
-                    </ComponentButton>
-                  )}
-
-                  {availableComponents.includes(COMPONENT_TYPE.SpawnPoint) && (
-                    <ComponentButton
-                      onClick={() => {
-                        const { engine } = useEditorStore.getState();
-                        if (!engine) return;
-
-                        const spawnPoint = engine.scene.extensions.spawn.createSpawnPoint();
-                        spawnPoint.title = SPAWN_TITLE.Default;
-                        node.setExtension("OMI_spawn_point", spawnPoint);
-                      }}
-                    >
-                      {COMPONENT_TYPE.SpawnPoint}
-                    </ComponentButton>
-                  )}
-                </div>
-              </DropdownContent>
-            </DropdownMenu>
-          </div>
+          <AddComponentButton
+            availableComponents={availableComponents}
+            node={node}
+            extras={extras}
+          />
         )}
       </div>
     </div>
-  );
-}
-
-function ComponentButton({ children, ...props }: DropdownMenuItemProps) {
-  return (
-    <DropdownItem
-      className="w-full cursor-default px-10 text-left outline-none focus:bg-neutral-200 active:opacity-80"
-      {...props}
-    >
-      {children}
-    </DropdownItem>
   );
 }
