+++ apps/client/src/editor/components/ui/NumberInput.tsx
@@ -1,12 +1,19 @@
+import { useEditorStore } from "../../store";
+
 type Props = React.InputHTMLAttributes<HTMLInputElement>;
 
 export default function NumberInput(props: Props) {
+  const isPlaying = useEditorStore((state) => state.isPlaying);
+
   return (
     <input
       {...props}
       type="number"
       inputMode="numeric"
-      className="w-full rounded border pl-1.5 leading-snug hover:bg-neutral-100 focus:bg-neutral-100"
+      disabled={isPlaying}
+      className={`w-full rounded border pl-1.5 leading-snug ${
+        isPlaying ? "" : "hover:bg-neutral-100 focus:bg-neutral-100"
+      } `}
     />
   );
 }
