+++ apps/client/src/editor/components/ui/TextInput.tsx
@@ -1,11 +1,18 @@
+import { useEditorStore } from "../../store";
+
 type Props = React.InputHTMLAttributes<HTMLInputElement>;
 
 export default function TextInput(props: Props) {
+  const isPlaying = useEditorStore((state) => state.isPlaying);
+
   return (
     <input
       {...props}
       type="text"
-      className="w-full rounded border pl-1.5 leading-snug hover:bg-neutral-100 focus:bg-neutral-100"
+      disabled={isPlaying}
+      className={`w-full rounded border pl-1.5 leading-snug ${
+        isPlaying ? "" : "hover:bg-neutral-100 focus:bg-neutral-100"
+      } `}
     />
   );
 }
